requirements.txt
discord.py==2.1.0
syllables==1.0.6